Delete old MenuSidebar file and rename new Sidebar to MenuSidebar
