+++ src/components/NavBar.svelte
@@ -1,20 +1,5 @@
 <script>
   import MenuSidebar from "./MenuSidebar.svelte";
-  import ArchiveSidebar from "./ArchiveSidebar.svelte";
-  import Sidebar from "./Sidebar.svelte";
-
-  let is_menu_sidebar_shown;
-  let is_archive_sidebar_shown;
-
-  function handleMenuButton() {
-    if (!is_menu_sidebar_shown && !is_archive_sidebar_shown) {
-      is_menu_sidebar_shown = !is_menu_sidebar_shown;
-    } else if (is_menu_sidebar_shown && !is_archive_sidebar_shown) {
-      is_menu_sidebar_shown = !is_menu_sidebar_shown;
-    } else if (!is_menu_sidebar_shown && is_archive_sidebar_shown) {
-      is_archive_sidebar_shown = !is_archive_sidebar_shown;
-    }
-  }
 </script>
 
 <div class="parent">
@@ -24,9 +9,7 @@
   <button class="scale-on-hover borderless-button">
     <img src="switch_board_view.png" alt="Board view" />
   </button>
-  <Sidebar />
-  <!-- <MenuSidebar bind:is_menu_sidebar_shown bind:is_archive_sidebar_shown />
-  <ArchiveSidebar bind:is_archive_sidebar_shown /> -->
+  <MenuSidebar />
 </div>
 
 <style>
